requirements.txt
requests
torch
ultralytics
datasets
scikit-learn
pillow
tqdm
numpy
imutils
supervision
